requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels-redis==3.3.1
charset-normalizer==2.0.7
click==8.0.3
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
Django==3.2.8
djangorestframework==3.12.4
gunicorn==20.1.0
h11==0.12.0
hiredis==2.0.0
httptools==0.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.2
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==21.0.0
python-dotenv==0.19.1
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.0
whitenoise==5.3.0
zope.interface==5.4.0
